Unit tests for suspendCoroutine withTimeout
